split block_concatenation.py
